Added rdnt-capital-locked-dlp strategy
